test: FreeSession 전용 테스트 확장 및 SessionTest 삭제
FreeSessionTest.java
- 공통 픽스처 setUp 추가
- 모집중 수강 신청 성공 테스트 추가
- 상태별(준비중/종료) 수강 신청 실패 테스트 추가
- 기존 생성자 검증 로직 유지 및 보강

SessionTest.java
- FreeSessionTest로 통합됨에 따라 클래스 삭제
