+++ src/test/java/nextstep/courses/domain/FreeSessionTest.java
@@ -3,18 +3,54 @@ package nextstep.courses.domain;
 import static org.assertj.core.api.Assertions.*;
 
 import java.time.LocalDate;
+import java.time.LocalDateTime;
+import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator;
 import org.junit.jupiter.api.Test;
+import org.junit.jupiter.params.ParameterizedTest;
+import org.junit.jupiter.params.provider.EnumSource;
 
 @DisplayNameGeneration(DisplayNameGenerator.ReplaceUnderscores.class)
 class FreeSessionTest {
 
+    private CoverImage coverImage;
+    private SessionPeriod period;
+
+    @BeforeEach
+    void setUp() {
+        coverImage = new CoverImage("image.png", 1024, 300, 200);
+        period = new SessionPeriod(LocalDate.of(2024, 1, 1), LocalDate.of(2024, 3, 31));
+    }
+
     @Test
     void 생성자_정상입력_생성성공() {
-        CoverImage coverImage = new CoverImage("image.png", 1024, 300, 200);
-        SessionPeriod sessionPeriod = new SessionPeriod(LocalDate.of(2024, 1, 1), LocalDate.of(2024, 3, 31));
+        FreeSession session = new FreeSession(coverImage, period);
 
-        assertThatCode(() -> new FreeSession(coverImage, sessionPeriod)).doesNotThrowAnyException();
+        assertThat(session.getStatus()).isEqualTo(SessionStatus.PREPARING);
+    }
+
+    @Test
+    void enroll_모집중_성공() {
+        Session session = new FreeSession(coverImage, period, SessionStatus.RECRUITING);
+        Enrollment enrollment = new Enrollment(1L, 1L, LocalDateTime.now());
+
+        session.enroll(enrollment, Money.ZERO);
+
+        assertThat(session.enrollmentCount()).isEqualTo(1);
+    }
+
+    @ParameterizedTest(name = "상태:{0}")
+    @EnumSource(
+            value = SessionStatus.class,
+            names = {"PREPARING", "CLOSED"})
+    void enroll_모집중이아닐시_예외발생(SessionStatus status) {
+        FreeSession session = new FreeSession(coverImage, period, status);
+
+        Enrollment enrollment = new Enrollment(1L, 1L, LocalDateTime.now());
+
+        assertThatThrownBy(() -> session.enroll(enrollment, Money.ZERO))
+                .isInstanceOf(IllegalStateException.class)
+                .hasMessageContaining("모집중인 강의만");
     }
 }
